Added accordion functionality to editor for a better representation
