+++ src/custom-blocks/accordion/index.js
@@ -125,98 +125,7 @@ registerBlockType("mojblocks/accordion-section", {
     }
 });
 
-/***
- * 
- * The following script is to mimic the behaviour of the accordion in the editor window.  
- * 
+/**
+ * Internal dependencies
  */
-
-
-function waitForElm(selector) {
-    return new Promise(resolve => {
-        if (document.querySelector(selector)) {
-            return resolve(document.querySelector(selector));
-        }
-
-        const observer = new MutationObserver(mutations => {
-            if (document.querySelector(selector)) {
-                resolve(document.querySelector(selector));
-                observer.disconnect();
-            }
-        });
-
-        observer.observe(document.body, {
-            childList: true,
-            subtree: true
-        });
-    });
-}
-
-
-const showThis_text = document.createTextNode("Show / Hide");
-const showThis = document.createElement("p");
-showThis.appendChild(showThis_text);
-showThis.classList.add("hideThis");
-showThis.classList.add("showHideThis");
-
-const showAll_text = document.createTextNode("Show all sections / Hide all sections");
-const showAll = document.createElement("p");
-showAll.appendChild(showAll_text);
-showAll.classList.add("hideAll");
-showAll.classList.add("showHideAll");
-
-waitForElm('.govuk-accordion').then((elm) => {
-    elm.insertBefore(showAll, elm.firstChild);
-
-    const sections = document.querySelectorAll(".govuk-accordion__section");
-
-    sections.forEach((section) => {
-        let showHideControl = showThis.cloneNode(true)
-        section.insertBefore(showHideControl, section.lastChild);
-        section.querySelector(".showHideThis").onclick = function() {
-            if (section.querySelector(".showHideThis").classList.contains("hideThis")) {
-                section.classList.add("accordion-hidden");
-                section.querySelector(".showHideThis").classList.add("showThis");
-                section.querySelector(".showHideThis").classList.remove("hideThis");
-                if (elm.getElementsByClassName("hideThis") == 0) {
-                    // if there is no hideThis left, we change the hide all to show all
-                    elm.querySelector(".showHideAll").classList.add("showAll");
-                    elm.querySelector(".showHideAll").classList.remove("hideAll");
-                }
-            } else {
-                section.classList.remove("accordion-hidden");
-                section.querySelector(".showHideThis").classList.add("hideThis");
-                section.querySelector(".showHideThis").classList.remove("showThis");
-                // there is now at least one hideThis, so we ensure hideAll is there instead of showAll
-                elm.querySelector(".showHideAll").classList.add("hideAll");
-                elm.querySelector(".showHideAll").classList.remove("showAll");
-            }
-        }
-    });
-
-    elm.querySelector(".showHideAll").onclick = function() {
-        if (elm.querySelector(".showHideAll").classList.contains("hideAll")) {
-            elm.querySelectorAll(".govuk-accordion__section").forEach(element => {
-                element.classList.add("accordion-hidden");
-                element.querySelectorAll(".showHideThis").forEach(element2 => {
-                    //we ensure any hideThis becomes a showThis
-                    element2.classList.add("showThis");
-                    element2.classList.remove("hideThis");
-                });
-            });
-            elm.querySelector(".showHideAll").classList.add("showAll");
-            elm.querySelector(".showHideAll").classList.remove("hideAll");
-        } else {
-            elm.querySelectorAll(".govuk-accordion__section").forEach(element => {
-                element.classList.remove("accordion-hidden");
-                element.querySelectorAll(".showHideThis").forEach(element2 => {
-                    //we ensure any showThis becomes a hideThis
-                    element2.classList.add("hideThis");
-                    element2.classList.remove("showThis");
-                });
-            });
-            elm.querySelector(".showHideAll").classList.remove("showAll");
-            elm.querySelector(".showHideAll").classList.add("hideAll");
-        }
-    };
-});
+import edit from './edit';
